Consolidated Treasury and vMooney analytics to one page w/ toggle
